third_party/blink/web_tests/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-public-expected.txt

This is a testharness.js-based test.
[FAIL] public to local: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] public to local: success.
  assert_equals: expected "loaded" but got "timeout"
[FAIL] public to private: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] public to private: success.
  assert_equals: expected "loaded" but got "timeout"
Harness: the test ran to completion.
